Moved project to Maven for dependency management and build automation
- Added Maven configuration files (pom.xml) to the project root.
- Reorganized project structure to conform to Maven conventions.
- Updated project dependencies to use Maven Central repositories.
- Updated build process to use Maven commands (clean, package).
- Updated README.md with instructions for building and running the project using Maven.
