+++ src/main/java/bits_please_api/APIRequestException.java
