+++ src/main/java/game/TrivialComputeGame.java
