+++ src/main/java/pages/GameSettingsPage.java
