+++ src/main/java/pages/QuestionEditorPage.java
@@ -140,7 +140,10 @@ public class QuestionEditorPage extends JFrame {
      */
     private void saveQuestions(){
         int rowCount = tableModel.getRowCount();
-        int currentQuestionsSize = tempQuestionList.size();
+        int currentQuestionsSize = 0;
+        if (tempQuestionList != null){
+            currentQuestionsSize = tempQuestionList.size();
+        }
         Object[][] questionsTableContent = parseAndValidate();
         if (questionsTableContent == null){
             return;
